Fixes for several bugs reported on trac
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@1176 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/kvirc/kvs/kvi_kvs_switchlist.cpp
@@ -57,7 +57,7 @@ void KviKvsSwitchList::addShort(unsigned short uShortKey,KviKvsVariant * pVarian
 {
 	if(!m_pShortSwitchDict)
 	{
-		m_pShortSwitchDict = new KviPointerHashTable<int,KviKvsVariant>(11);
+		m_pShortSwitchDict = new KviPointerHashTable<unsigned short,KviKvsVariant>(11);
 		m_pShortSwitchDict->setAutoDelete(true);
 	}
 	m_pShortSwitchDict->replace(uShortKey,pVariant);
